boundary detection with crop classification/requirements.txt
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2
google-generativeai==0.3.2
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
python-multipart==0.0.6
